refactor: restructure buildkit into modular files for improved maintainability
- Created builder.js to handle build and development commands.
- Added help.js to provide a help message for the CLI.
- Implemented test.js to run tests using Node.js test runner.
- Updated index.js to serve as the main entry point for the CLI, routing commands to their respective modules.
- Introduced TypeScript declaration generation and improved logging throughout the build process.
