+++ .changeset/silent-dogs-wash.md
@@ -1,0 +1,5 @@
+---
+"@withstudiocms/buildkit": patch
+---
+
+Refactors buildkit to be broken out into individual files to help with maintainability
